requirements.txt
torch
torchvision
numpy
pandas
opencv-python-headless
scikit-learn
matplotlib
umap-learn
streamlit>=1.31.0
altair<5
geopandas
folium
streamlit-folium
pyyaml
pillow
tqdm
scipy
plotly
